+++ src/wildboar/linear_model/_shapelet.py
@@ -1,7 +1,7 @@
 # Authors: Isak Samsten
 # License: BSD 3 clause
 
-from ..transform import RandomShapeletTransform
+from ..transform import DilatedShapeletTransform, RandomShapeletTransform
 from ._transform import TransformRidgeClassifierCV, TransformRidgeCV
 
 
@@ -23,7 +23,7 @@ class RandomShapeletClassifier(TransformRidgeClassifierCV):
         cv=None,
         class_weight=None,
         n_jobs=None,
-        random_state=None
+        random_state=None,
     ):
         super().__init__(
             alphas=alphas,
@@ -71,7 +71,7 @@ class RandomShapeletRegressor(TransformRidgeCV):
         cv=None,
         gcv_mode=None,
         n_jobs=None,
-        random_state=None
+        random_state=None,
     ):
         super().__init__(
             alphas=alphas,
@@ -96,6 +96,128 @@ class RandomShapeletRegressor(TransformRidgeCV):
             metric_params=self.metric_params,
             min_shapelet_size=self.min_shapelet_size,
             max_shapelet_size=self.max_shapelet_size,
+            random_state=random_state,
+            n_jobs=self.n_jobs,
+        )
+
+
+class DilatedShapeletClassifier(TransformRidgeClassifierCV):
+    """
+    A classifier that uses random dilated shapelets.
+
+    Parameters
+    ----------
+    n_shapelets : int, optional
+        The number of dilated shapelets.
+    metric : str or callable, optional
+        The distance metric
+
+        See ``_METRICS.keys()`` for a list of supported metrics.
+    metric_params : dict, optional
+        Parameters to the metric.
+
+        Read more about the parameters in the
+        :ref:`User guide <list_of_metrics>`.
+    normalize_prob : float, optional
+        The probability of standardizing a shapelet with zero mean and unit
+        standard deviation.
+    min_shapelet_size : float, optional
+        The minimum shapelet size. If None, use the discrete sizes
+        in `shapelet_size`.
+    max_shapelet_size : float, optional
+        The maximum shapelet size. If None, use the discrete sizes
+        in `shapelet_size`.
+    shapelet_size : array-like, optional
+        The size of shapelets.
+    lower : float, optional
+        The lower percentile to draw distance thresholds above.
+    upper : float, optional
+        The upper percentile to draw distance thresholds below.
+    alphas : array-like of shape (n_alphas,), optional
+        Array of alpha values to try.
+    fit_intercept : bool, optional
+        Whether to calculate the intercept for this model.
+    scoring : str, callable, optional
+        A string or a scorer callable object with signature
+        `scorer(estimator, X, y)`.
+    cv : int, cross-validation generator or an iterable, optional
+        Determines the cross-validation splitting strategy.
+    class_weight : dict or 'balanced', optional
+        Weights associated with classes in the form `{class_label: weight}`.
+    normalize : bool, optional
+        Standardize before fitting.
+    random_state : int or RandomState, optional
+        Controls the random sampling of kernels.
+
+        - If `int`, `random_state` is the seed used by the random number
+          generator.
+        - If :class:`numpy.random.RandomState` instance, `random_state` is
+          the random number generator.
+        - If `None`, the random number generator is the
+          :class:`numpy.random.RandomState` instance used by
+          :func:`numpy.random`.
+    n_jobs : int, optional
+        The number of parallel jobs.
+
+    References
+    ----------
+    Antoine Guillaume, Christel Vrain, Elloumi Wael
+        Random Dilated Shapelet Transform: A New Approach for Time Series Shapelets
+        Pattern Recognition and Artificial Intelligence, 2022
+    """
+
+    def __init__(  # noqa: PLR0913
+        self,
+        n_shapelets=1000,
+        *,
+        metric="euclidean",
+        metric_params=None,
+        normalize_prob=0.8,
+        min_shapelet_size=None,
+        max_shapelet_size=None,
+        shapelet_size=None,
+        lower=0.05,
+        upper=0.1,
+        alphas=(0.1, 1.0, 10.0),
+        fit_intercept=True,
+        scoring=None,
+        cv=None,
+        class_weight=None,
+        normalize=False,
+        random_state=None,
+        n_jobs=None,
+    ):
+        super().__init__(
+            alphas=alphas,
+            fit_intercept=fit_intercept,
+            normalize=normalize,
+            scoring=scoring,
+            cv=cv,
+            class_weight=class_weight,
+            n_jobs=n_jobs,
+            random_state=random_state,
+        )
+        self.n_shapelets = n_shapelets
+        self.metric = metric
+        self.metric_params = metric_params
+        self.shapelet_size = shapelet_size
+        self.min_shapelet_size = min_shapelet_size
+        self.max_shapelet_size = max_shapelet_size
+        self.normalize_prob = normalize_prob
+        self.lower = lower
+        self.upper = upper
+
+    def _get_transform(self, random_state):
+        return DilatedShapeletTransform(
+            n_shapelets=self.n_shapelets,
+            metric=self.metric,
+            metric_params=self.metric_params,
+            shapelet_size=self.shapelet_size,
+            normalize_prob=self.normalize_prob,
+            min_shapelet_size=self.min_shapelet_size,
+            max_shapelet_size=self.max_shapelet_size,
+            lower=self.lower,
+            upper=self.upper,
             random_state=random_state,
             n_jobs=self.n_jobs,
         )